requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Markdown==3.3.4
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
